[IDLE-179] feat: 마이북 목록/상세 UI 구현 및 api 연동 - 마이북 목록 및 상세 UI 구현 - 마이북 목록 조회 api 연동 - 마이북 상세 조회 api 연동

+++ frontend/lib/data/model/book/mybooks_response.dart
@@ -79,7 +79,7 @@ class BookInfo {
   String? title;
   String? description;
   String? thumbnailUrl;
-  int? stars;
+  double? stars;
 
   BookInfo({
     this.id,
